LayoutTests/imported/w3c/web-platform-tests/encoding/textdecoder-utf16-surrogates-expected.txt


FAIL utf-16le - lone surrogate lead assert_equals: expected "\ufffd" but got "í "
FAIL utf-16le - lone surrogate lead (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fatal: true}).decode(new Uint8Array(t.input))
        }" did not throw
FAIL utf-16le - lone surrogate trail assert_equals: expected "\ufffd" but got "í°"
FAIL utf-16le - lone surrogate trail (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fatal: true}).decode(new Uint8Array(t.input))
        }" did not throw
FAIL utf-16le - unmatched surrogate lead assert_equals: expected "\ufffd\0" but got "í \0"
FAIL utf-16le - unmatched surrogate lead (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fatal: true}).decode(new Uint8Array(t.input))
        }" did not throw
FAIL utf-16le - unmatched surrogate trail assert_equals: expected "\ufffd\0" but got "í°\0"
FAIL utf-16le - unmatched surrogate trail (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fatal: true}).decode(new Uint8Array(t.input))
        }" did not throw
FAIL utf-16le - swapped surrogate pair assert_equals: expected "\ufffd\ufffd" but got "í°í "
FAIL utf-16le - swapped surrogate pair (fatal flag set) assert_throws: function "function () {
            new TextDecoder(t.encoding, {fatal: true}).decode(new Uint8Array(t.input))
        }" did not throw
